+++ packages/dagonizer-embedder-mistral/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @noocodex/dagonizer-embedder-mistral
 
+## 0.17.0
+
 ## 0.16.0
 
 ## 0.15.0